# apps/api/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AccessController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "operationId": "AccessController_password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "AccessController_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      },
      "post": {
        "operationId": "AccessController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
    "/api/v1/users/access": {
      "post": {
        "operationId": "AccessController_change",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
    "/api/v1/users/remove": {
      "post": {
        "operationId": "AccessController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
...
  },
  "info": {
    "title": "IOP local API host",
    "description": "Health plus explicitly activated local analytical POC. Business operations require a configured local demo session, current site permissions and exact configured scope; no shared-user authentication claim.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "SuccessDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "PasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "writeOnly": true
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 15,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "password"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "profile": {
            "type": "string",
            "enum": [
              "administrator",
              "technician",
              "task-force",
              "team-leader"
            ]
          },
          "id": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "username",
          "profile",
          "id",
          "active"
        ]
      },
      "NewUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "profile": {
            "type": "string",
            "enum": [
              "administrator",
              "technician",
              "task-force",
              "team-leader"
            ]
          }
        },
        "required": [
          "name",
          "username",
          "profile"
        ]
      },
      "CreatedUserDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserProfileDto"
          },
          "initialPassword": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "initialPassword"
        ]
      },
      "ChangeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "profile": {
            "type": "string",
            "enum": [
              "administrator",
              "technician",
              "task-force",
              "team-leader"
            ]
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "profile",
          "active"
        ]
      },
      "RemoveUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}